Update: Updated the confirm model function.

+++ services/classification.py
@@ -67,25 +67,61 @@ class ClassificationService:
         return prediction[0]
 
     @staticmethod
+    # def confirm(content, label, model='public'):
+    #     try:
+    #         corpus = []
+    #         corpus.append(content)
+    #         embeddings = embedder.encode(corpus)
+    #         with open('{PATH}/model/model_{schema_name}.pkl'.format(schema_name=model, PATH=PATH), 'rb') as file:
+    #             learner = pickle.load(file)
+    #         if isinstance(label, str):
+    #             label = numpy.array([label])
+    #         elif isinstance(label, list):
+    #             label = numpy.array(label)
+    #             pass
+    #         else:
+    #             raise ValueError("label must be a string or a list")
+    #         print(numpy.array([label]))
+    #         learner.teach(embeddings.reshape(1, -1), numpy.array([label]))
+    #         with open('{PATH}/model/model_{schema_name}.pkl'.format(schema_name=model, PATH=PATH), 'wb') as file:
+    #             pickle.dump(learner, file, protocol=pickle.HIGHEST_PROTOCOL)
+    #             print('Model Saved!')
+    #     except Exception as e:
+    #         print(e)
+    #     return True
     def confirm(content, label, model='public'):
         try:
             corpus = []
             corpus.append(content)
             embeddings = embedder.encode(corpus)
-            with open('{PATH}/model/model_{schema_name}.pkl'.format(schema_name=model, PATH=PATH), 'rb') as file:
+
+            model_path = '{PATH}/model/model_{schema_name}.pkl'.format(
+                schema_name=model, PATH=PATH)
+
+            # Check if the file exists
+            if not exists(model_path):
+                raise FileNotFoundError(f"File {model_path} not found!")
+
+            # Read the pickle file
+            with open(model_path, 'rb') as file:
                 learner = pickle.load(file)
+
             if isinstance(label, str):
                 label = numpy.array([label])
             elif isinstance(label, list):
                 label = numpy.array(label)
-                pass
             else:
                 raise ValueError("label must be a string or a list")
+
             print(numpy.array([label]))
             learner.teach(embeddings.reshape(1, -1), numpy.array([label]))
-            with open('{PATH}/model/model_{schema_name}.pkl'.format(schema_name=model, PATH=PATH), 'wb') as file:
+
+            # Write the pickle file
+            with open(model_path, 'wb') as file:
                 pickle.dump(learner, file, protocol=pickle.HIGHEST_PROTOCOL)
                 print('Model Saved!')
+
         except Exception as e:
             print(e)
+
         return True
